requirements.txt
fastapi
uvicorn
deta
python-multipart
python-dotenv
jinja2
pyjwt
